refactor(invoke): split route adapters and stream transports (#840)
